Restructure imports and exports
